src/function_app/requirements.txt
azure-functions>=1.21.0
azure-eventhub>=5.12.0
azure-identity>=1.17.1
azure-storage-file-datalake>=12.16.0
pymongo>=4.10.0
